move answer logic from app.py
