+++ backend/app.py
@@ -1,10 +1,11 @@
 import os
-from typing import Any
 
 from dotenv import load_dotenv
 from fastapi import FastAPI, Header, HTTPException
 from fastapi.middleware.cors import CORSMiddleware
-from pydantic import BaseModel
+
+from killtherobot_backend.llms import get_answers
+from killtherobot_backend.models import Answer, Game
 
 _ = load_dotenv()
 
@@ -24,40 +25,6 @@ app.add_middleware(
 @app.get("/healthz")
 async def healthz():
     return {"status": "healthy"}
-
-
-class Human(BaseModel):
-    name: str
-    isAlive: bool
-    isAdmin: bool
-
-
-class Bot(BaseModel):
-    name: str
-    config: Any
-    isAlive: bool
-
-
-class Answer(BaseModel):
-    name: str
-    text: str
-    votes: list[str] | None
-
-
-class Round(BaseModel):
-    question: str
-    answers: list[Answer]
-
-
-class Game(BaseModel):
-    name: str
-    theme: str
-    code: str
-    stage: str
-    humans: list[Human]
-    bots: list[Bot]
-    rounds: list[Round]
-    currentRound: int
 
 
 @app.post("/get_answer")
@@ -80,8 +47,6 @@ async def answer_question(game: Game, api_key: str = Header(...)) -> list[Answer
     question = game.rounds[-1].question
     names = [bot.name for bot in game.bots]
 
-    # TODO: Implement the logic to answer the question using OpenAI API
-    # This is a placeholder response
-    return [
-        Answer(name=name, text=f"answer to {question}", votes=None) for name in names
-    ]
+    answers = get_answers(question, names)
+
+    return answers
